+++ apps/aurora-portal/src/client/routes/_auth/accounts/index.tsx
@@ -1,0 +1,14 @@
+import { createFileRoute, redirect } from "@tanstack/react-router"
+
+export const Route = createFileRoute("/_auth/accounts/")({
+  component: RouteComponent,
+  beforeLoad: ({ context }) => {
+    if (context.auth?.isAuthenticated) {
+      throw redirect({ to: "/auth/login" })
+    }
+  },
+})
+
+function RouteComponent() {
+  return <div>Hello "/accounts/"!</div>
+}
